Implemented Admin API
GetSalesReport relies on quantity to change for the math to work, which can only happen when the transaction is processed which will update quantity to the number of remaining vehicles.  total sold = sum of (initial quantity - current quantity) for all vehicles. As such checkout should update quantity while initial quantity will remain the same.